+++ locadora.diurno.dal/src/locadora/diurno/dal/dao/interfaces/IAutomovelDAO.java
@@ -1,8 +1,9 @@
 package locadora.diurno.dal.dao.interfaces;
 
-import locadora.diurno.dal.generics.*;		
+import locadora.diurno.dal.generics.*;
 import locadora.diurno.dal.entidade.*;
 
-public interface IAutomovelDAO extends IGenericDAO<Automovel, Short>{
+public interface IAutomovelDAO
+		extends IGenericDAO<Automovel,Integer>{
 
 }
